Split scribe.py: extract resolve.py + move _ProgressTally to progress.py
Phase-0 refactor (#306), commit 1 of 5. Behaviour-preserving extraction —
moves only, no logic changes.

- New bartleby/ingest/resolve.py: the provider resolvers
  (_resolve_{llm,vision}_provider), the worker-count resolvers
  (_resolve_{max,caption,summarize}_workers), and _required_hf_models.
- _ProgressTally moves into bartleby/ingest/progress.py (its callers
  _caption_all/_summarize_all import it from there for now).
- commands/scribe.py imports `from bartleby.ingest import resolve` and calls
  resolve._resolve_*; drops the now-unused imports (os, psutil, get_provider,
  ensure_provider_env, and the worker/HF consts).
- Tests retargeted to the new namespaces: the resolver unit tests alias
  bartleby.ingest.resolve, get_provider patches point at resolve.get_provider,
  and test_quiet imports _required_hf_models from resolve.
